style: move flowbite offcanvas customTheme setting to common page
