+++ src/components/landLordManagement/request/RequestCard.tsx
@@ -1,6 +1,5 @@
 import { useState } from "react";
 import { Tooltip, Drawer } from "flowbite-react";
-import { CustomFlowbiteTheme, Flowbite } from "flowbite-react";
 import rightIcon_black from "../../../assets/imgs/icons/rightIcon_black.svg";
 import close from "../../../assets/imgs/icons/close.svg";
 import photo from "../../../assets/imgs/homePage/recommendation_picture_1.svg"
@@ -9,36 +8,6 @@ import star from "../../../assets/imgs/icons/star.svg";
 export default function RequestCard({status = "none"}) {
   const [isOpen, setIsOpen] = useState(false);
   const handleClose = () => setIsOpen(false);
-  const customTheme: CustomFlowbiteTheme = {
-    drawer: {
-      "root": {
-        "base": "fixed z-40 overflow-y-auto bg-Landlord-95 px-10 pt-10 transition-transform",
-        "backdrop": "fixed inset-0 z-30 bg-transparent",
-        "edge": "bottom-16",
-        "position": {
-          "right": {
-            "on": "shadow-elevation-3 right-0 top-[64px] bottom-0 w-5/12 transform-none scrollbar-hide",
-            "off": "right-0 top-[152px] h-screen w-5/12 translate-x-full"
-          },
-        }
-      },
-      "header": {
-        "inner": {
-          "closeButton": "absolute end-2.5 top-2.5 flex h-8 w-8 items-center justify-center rounded-lg bg-transparent text-sm text-gray-400 hover:bg-gray-200 hover:text-gray-900 dark:hover:bg-gray-600 dark:hover:text-white",
-          "closeIcon": "h-4 w-4",
-          "titleIcon": "me-2.5 h-4 w-4",
-          "titleText": "mb-4 inline-flex items-center text-base font-semibold text-gray-500 dark:text-gray-400"
-        },
-        "collapsed": {
-          "on": "hidden",
-          "off": "block"
-        }
-      },
-      "items": {
-        "base": ""
-      }
-    }
-  };
 
   return (
     <>
@@ -91,163 +60,85 @@ export default function RequestCard({status = "none"}) {
           <button
             type="button"
             className="mt-auto mb-1 mr-2 letter-button-light border-b border-black hover:border-Neutral-30"
-            // onClick={() => setIsOpen(true)}
           >
             查看更多
             <img src={rightIcon_black} alt="rightIcon" />
           </button>
-          {/* <Flowbite theme={{ theme: customTheme }}>
-            <Drawer open={isOpen} onClose={handleClose} position="right">
-              <Drawer.Items>
-                <div className="layout-grid gap-4 mb-32">
-                  <div className="col-span-10 flex flex-col gap-6 items-end">
-                    <button
-                      type="button"
-                      onClick={handleClose}
-                    >
-                      <img src={close} alt="close" />
-                    </button>
-                    <button
-                      type="button"
-                      className="outline-button-m flex items-center gap-2"
-                    >
-                      隱藏租客
-                      <span className="material-symbols-outlined text-base">visibility_off</span>
-                    </button>
-                    <div className="w-full p-6 rounded-2xl bg-white">
-                      <div className="flex justify-between mb-3.5">
-                        <div>
-                          <span className="material-symbols-outlined text-Tenant-50 mb-6">face</span>
-                          <h4 className="text-sans-b-body1 text-Tenant-50 mb-3">租客</h4>
-                          <h3 className="text-sans-b-h4 mb-3">詹小美</h3>
-                          <div className="text-sans-body1 flex mb-2">
-                            <h4 className="pr-2 mr-2 border-r border-Tenant-70">女</h4>
-                            <h4>文教類</h4>
-                          </div>
-                          <a href="tel:+886-958-230-1239" className="text-sans-body1">0958-230-1239</a>
-                        </div>
-                        <div className="w-[186px] rounded-lg overflow-hidden">
-                          <img src={photo} alt="tenant" className="w-full aspect-square block object-cover" />
-                        </div>
-                      </div>
-                      <div className="flex gap-3 mb-6">
-                        <div className="flex-1 p-4 rounded-lg bg-Neutral-95 flex gap-4 flex-wrap justify-between">
-                          <h5 className="w-full text-Tenant-50 text-sans-b-body1">評價</h5>
-                          <h4 className="text-sans-h4">4.8</h4>
-                          <img src={star} alt="star" className="self-end" />
-                        </div>
-                        <div className="flex-1 p-4 rounded-lg bg-Neutral-95 flex gap-4 flex-wrap justify-between">
-                          <h5 className="w-full text-Tenant-50 text-sans-b-body1">則數</h5>
-                          <h4 className="text-sans-h4">3</h4>
-                          <h6 className="self-end">則</h6>
-                        </div>
-                      </div>
-                      <div>
-                        <h4 className="text-Landlord-40 text-sans-b-body1 mb-3">自我介紹</h4>
-                        <p className="text-sans-body1">我是來自台中的補習班老師。最近決定轉職成為前端工程師，於是報名參加了高雄六角的培訓營！我希望能在開課前找到便宜、便利，而且能申請租屋補助的房源，同時也希望能遇到一位優質的房東。</p>
-                      </div>
-                    </div>
-                    <div className="w-full">
-                      <h4 className="text-sans-b-h6 mb-4">房東評價</h4>
-                      <div className="bg-white rounded-lg p-4">
-                        <div className="flex flex-col gap-2">
-                          <h5 className="text-sans-b-body1 text-Landlord-50">王太太</h5>
-                          <time className="text-sans-body1">2023年4月22日</time>
-                          <div className="flex gap-2">
-                            <img src={star} alt="star" />
-                            <img src={star} alt="star" />
-                            <img src={star} alt="star" />
-                          </div>
-                          <p className="mt-2 text-sans-body1">黃同學乾淨有禮貌，好租客！</p>
-                        </div>
-                        <div className="mt-4 pt-6 border-t border-Neutral-95">
-                          <h5 className="text-sans-b-body1 text-Landlord-50 mb-4">租客回覆</h5>
-                          <p className="mt-2 text-sans-body1">謝謝王媽媽照顧！</p>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </Drawer.Items>
-            </Drawer>
-          </Flowbite> */}
         </div>
       </li>
-      <Flowbite theme={{ theme: customTheme }}>
-        <Drawer open={isOpen} onClose={handleClose} position="right">
-          <Drawer.Items>
-            <div className="layout-grid gap-4 mb-32">
-              <div className="col-span-10 flex flex-col gap-6 items-end">
-                <button
-                  type="button"
-                  onClick={handleClose}
-                >
-                  <img src={close} alt="close" />
-                </button>
-                <button
-                  type="button"
-                  className="outline-button-m flex items-center gap-2"
-                >
-                  隱藏租客
-                  <span className="material-symbols-outlined text-base">visibility_off</span>
-                </button>
-                <div className="w-full p-6 rounded-2xl bg-white">
-                  <div className="flex justify-between mb-3.5">
-                    <div>
-                      <span className="material-symbols-outlined text-Tenant-50 mb-6">face</span>
-                      <h4 className="text-sans-b-body1 text-Tenant-50 mb-3">租客</h4>
-                      <h3 className="text-sans-b-h4 mb-3">詹小美</h3>
-                      <div className="text-sans-body1 flex mb-2">
-                        <h4 className="pr-2 mr-2 border-r border-Tenant-70">女</h4>
-                        <h4>文教類</h4>
-                      </div>
-                      <a href="tel:+886-958-230-1239" className="text-sans-body1">0958-230-1239</a>
-                    </div>
-                    <div className="w-[186px] rounded-lg overflow-hidden">
-                      <img src={photo} alt="tenant" className="w-full aspect-square block object-cover" />
-                    </div>
-                  </div>
-                  <div className="flex gap-3 mb-6">
-                    <div className="flex-1 p-4 rounded-lg bg-Neutral-95 flex gap-4 flex-wrap justify-between">
-                      <h5 className="w-full text-Tenant-50 text-sans-b-body1">評價</h5>
-                      <h4 className="text-sans-h4">4.8</h4>
-                      <img src={star} alt="star" className="self-end" />
-                    </div>
-                    <div className="flex-1 p-4 rounded-lg bg-Neutral-95 flex gap-4 flex-wrap justify-between">
-                      <h5 className="w-full text-Tenant-50 text-sans-b-body1">則數</h5>
-                      <h4 className="text-sans-h4">3</h4>
-                      <h6 className="self-end">則</h6>
-                    </div>
-                  </div>
+      <Drawer open={isOpen} onClose={handleClose} position="right">
+        <Drawer.Items>
+          <div className="layout-grid gap-4 mb-32">
+            <div className="col-span-10 flex flex-col gap-6 items-end">
+              <button
+                type="button"
+                onClick={handleClose}
+              >
+                <img src={close} alt="close" />
+              </button>
+              <button
+                type="button"
+                className="outline-button-m flex items-center gap-2"
+              >
+                隱藏租客
+                <span className="material-symbols-outlined text-base">visibility_off</span>
+              </button>
+              <div className="w-full p-6 rounded-2xl bg-white">
+                <div className="flex justify-between mb-3.5">
                   <div>
-                    <h4 className="text-Landlord-40 text-sans-b-body1 mb-3">自我介紹</h4>
-                    <p className="text-sans-body1">我是來自台中的補習班老師。最近決定轉職成為前端工程師，於是報名參加了高雄六角的培訓營！我希望能在開課前找到便宜、便利，而且能申請租屋補助的房源，同時也希望能遇到一位優質的房東。</p>
+                    <span className="material-symbols-outlined text-Tenant-50 mb-6">face</span>
+                    <h4 className="text-sans-b-body1 text-Tenant-50 mb-3">租客</h4>
+                    <h3 className="text-sans-b-h4 mb-3">詹小美</h3>
+                    <div className="text-sans-body1 flex mb-2">
+                      <h4 className="pr-2 mr-2 border-r border-Tenant-70">女</h4>
+                      <h4>文教類</h4>
+                    </div>
+                    <a href="tel:+886-958-230-1239" className="text-sans-body1">0958-230-1239</a>
+                  </div>
+                  <div className="w-[186px] rounded-lg overflow-hidden">
+                    <img src={photo} alt="tenant" className="w-full aspect-square block object-cover" />
                   </div>
                 </div>
-                <div className="w-full">
-                  <h4 className="text-sans-b-h6 mb-4">房東評價</h4>
-                  <div className="bg-white rounded-lg p-4">
-                    <div className="flex flex-col gap-2">
-                      <h5 className="text-sans-b-body1 text-Landlord-50">王太太</h5>
-                      <time className="text-sans-body1">2023年4月22日</time>
-                      <div className="flex gap-2">
-                        <img src={star} alt="star" />
-                        <img src={star} alt="star" />
-                        <img src={star} alt="star" />
-                      </div>
-                      <p className="mt-2 text-sans-body1">黃同學乾淨有禮貌，好租客！</p>
+                <div className="flex gap-3 mb-6">
+                  <div className="flex-1 p-4 rounded-lg bg-Neutral-95 flex gap-4 flex-wrap justify-between">
+                    <h5 className="w-full text-Tenant-50 text-sans-b-body1">評價</h5>
+                    <h4 className="text-sans-h4">4.8</h4>
+                    <img src={star} alt="star" className="self-end" />
+                  </div>
+                  <div className="flex-1 p-4 rounded-lg bg-Neutral-95 flex gap-4 flex-wrap justify-between">
+                    <h5 className="w-full text-Tenant-50 text-sans-b-body1">則數</h5>
+                    <h4 className="text-sans-h4">3</h4>
+                    <h6 className="self-end">則</h6>
+                  </div>
+                </div>
+                <div>
+                  <h4 className="text-Landlord-40 text-sans-b-body1 mb-3">自我介紹</h4>
+                  <p className="text-sans-body1">我是來自台中的補習班老師。最近決定轉職成為前端工程師，於是報名參加了高雄六角的培訓營！我希望能在開課前找到便宜、便利，而且能申請租屋補助的房源，同時也希望能遇到一位優質的房東。</p>
+                </div>
+              </div>
+              <div className="w-full">
+                <h4 className="text-sans-b-h6 mb-4">房東評價</h4>
+                <div className="bg-white rounded-lg p-4">
+                  <div className="flex flex-col gap-2">
+                    <h5 className="text-sans-b-body1 text-Landlord-50">王太太</h5>
+                    <time className="text-sans-body1">2023年4月22日</time>
+                    <div className="flex gap-2">
+                      <img src={star} alt="star" />
+                      <img src={star} alt="star" />
+                      <img src={star} alt="star" />
                     </div>
-                    <div className="mt-4 pt-6 border-t border-Neutral-95">
-                      <h5 className="text-sans-b-body1 text-Landlord-50 mb-4">租客回覆</h5>
-                      <p className="mt-2 text-sans-body1">謝謝王媽媽照顧！</p>
-                    </div>
+                    <p className="mt-2 text-sans-body1">黃同學乾淨有禮貌，好租客！</p>
+                  </div>
+                  <div className="mt-4 pt-6 border-t border-Neutral-95">
+                    <h5 className="text-sans-b-body1 text-Landlord-50 mb-4">租客回覆</h5>
+                    <p className="mt-2 text-sans-body1">謝謝王媽媽照顧！</p>
                   </div>
                 </div>
               </div>
             </div>
-          </Drawer.Items>
-        </Drawer>
-      </Flowbite>
+          </div>
+        </Drawer.Items>
+      </Drawer>
     </>
   );
 }
